Moving this function to node.py
