+++ ocs_ci/ocs/cluster.py
@@ -18,7 +18,6 @@ import ocs_ci.ocs.resources.pod as pod
 from ocs_ci.ocs.exceptions import UnexpectedBehaviour
 from ocs_ci.ocs.resources import ocs, storage_cluster
 import ocs_ci.ocs.constants as constant
-from ocs_ci.ocs.node import get_typed_nodes
 from ocs_ci.utility.retry import retry
 from ocs_ci.utility.utils import TimeoutSampler, run_cmd
 from ocs_ci.ocs.utils import get_pod_name_by_pattern
@@ -744,43 +743,3 @@ def validate_osd_utilization(osd_used=80):
             logger.warn(f"{osd} used value {value}")
 
     return _rc
-
-
-def get_node_resource_utilization(nodename=None, node_type='worker'):
-    """
-    Gets the node's cpu and memory utilization in percentage
-
-    Args:
-        nodename (str) : The node name
-        node_type (str) : The node type (e.g. master, worker)
-
-    Returns:
-        dict : Node name and its cpu and memory utilization in
-               percentage
-
-    """
-
-    node_names = [nodename] if nodename else [
-        node.name for node in get_typed_nodes(node_type=node_type)
-    ]
-    obj = ocp.OCP()
-    resource_utilization_all_nodes = obj.exec_oc_cmd(
-        command='adm top nodes', out_yaml_format=False
-    ).split("\n")
-    utilization_dict = {}
-
-    for node in node_names:
-        for value in resource_utilization_all_nodes:
-            if node in value:
-                value = re.findall(r'\d+', value.strip())
-                cpu_utilization = value[2]
-                logger.info("The CPU utilized by the node "
-                            f"{node} is {cpu_utilization}%")
-                memory_utilization = value[4]
-                logger.info("The memory utilized of the node "
-                            f"{node} is {memory_utilization}%")
-                utilization_dict[node] = {
-                    'cpu': int(cpu_utilization),
-                    'memory': int(memory_utilization)
-                }
-    return utilization_dict
